services/dashboard_api/requirements.txt
flask
psycopg2-binary
flask-cors
prometheus-client
